ME-460 According to the user authentication pages should load and Display the app preview and create a new app

+++ assets/js/controllers/welcome/livePreviewCtrl.js
@@ -9,12 +9,12 @@
     'use strict';
     angular.module('app')
         .controller('livePreviewCtrl',
-        ['$scope', 'welcomeTemplatesResource', 'userProfileService','$stateParams','$mdDialog','mySharedService','$http','$state',
+        ['$scope', 'welcomeTemplatesResource', 'userProfileService','$stateParams','$mdDialog','mySharedService','$http','$state','$auth',
             livePreviewCtrl
         ]);
 
 
-    function livePreviewCtrl($scope,welcomeTemplatesResource, userProfileService,$stateParams,$mdDialog,mySharedService,$http,$state) {
+    function livePreviewCtrl($scope,welcomeTemplatesResource, userProfileService,$stateParams,$mdDialog,mySharedService,$http,$state,$auth) {
         welcomeTemplatesResource.getTemplates().success(function(data){
             $scope.templates = data;
         });
@@ -33,38 +33,42 @@
             };
 
             welcomeTemplatesResource.deletePreviewTemp(appParams).then(function(data) {
-                $state.go('user.welcome');
+                $state.go('anon.welcome');
             });
         };
 
         $scope.open = function(templateId, templateUrl, templateName) {
+            if ($auth.isAuthenticated()) {
+                $scope.deleteFile($stateParams.userId, $stateParams.appId);
 
-            $scope.deleteFile($stateParams.userId, $stateParams.appId);
-
-            $mdDialog.show({
-                controller: DialogController,
-                templateUrl: 'user/welcome/startAppWindowView.html',
-                parent: angular.element(document.body),
-                //targetEvent: ev,
-                resolve: {
-                    initialData:["$q",function($q){
-                        return $q.all({
-                            selectedTemplateUrl: templateUrl,
-                            selectedTemplateId : templateId,
-                            selectedTemplateName:  templateName
-                        })
-                    }]
-                },
-                clickOutsideToClose:true
-            })
-                .then(function(answer) {
-                    $scope.status = 'You said the information was "' + answer + '".';
-                }, function() {
-                    $scope.status = 'You cancelled the dialog.';
-                });
-        };
-        $scope.profileView = function() {
-            return userProfileService.showUserProfileDialog();
+                $mdDialog.show({
+                    controller: DialogController,
+                    templateUrl: 'user/welcome/startAppWindowView.html',
+                    parent: angular.element(document.body),
+                    //targetEvent: ev,
+                    resolve: {
+                        initialData: ["$q", function ($q) {
+                            return $q.all({
+                                selectedTemplateUrl: templateUrl,
+                                selectedTemplateId: templateId,
+                                selectedTemplateName: templateName
+                            })
+                        }]
+                    },
+                    clickOutsideToClose: true
+                })
+                    .then(function (answer) {
+                        $scope.status = 'You said the information was "' + answer + '".';
+                    }, function () {
+                        $scope.status = 'You cancelled the dialog.';
+                    });
+            }else{
+                $state.go('anon.login');
+            }
+            ;
+            $scope.profileView = function () {
+                return userProfileService.showUserProfileDialog();
+            }
         }
     }
 
@@ -87,17 +91,15 @@
 
         $scope.answer = function(answer) {
 
-            var appParams = {
+            var tempAppParams = {
                 'appName': $scope.appName,
                 'templateId': initialData.selectedTemplateId,
                 'templateName': initialData.selectedTemplateName,
-                'templateUrl':initialData.selectedTemplateUrl
+                'templateUrl':initialData.selectedTemplateUrl,
+                'userId':$auth.getPayload().id
 
             };
-            console.log(appParams);
-            if ($auth.isAuthenticated()) {
-
-                welcomeTemplatesResource.createApp(appParams).then(function(data){
+                welcomeTemplatesResource.createApp(tempAppParams).then(function(data){
 
                     var url= ME_APP_SERVER+'temp/'+$auth.getPayload().id
                         +'/templates/'+data.data.appId+'/?'+new Date().getTime();
@@ -105,9 +107,6 @@
                     mySharedService.prepForBroadcast(url);
                     $state.go('user.editApp',{appId:data.data.appId});
                 });
-            }else{
-                $state.go('anon.login');
-            }
             $mdDialog.hide(answer);
         };
     };
